working on Voter Info. Can parse voterinfo response, at least obejct is created

+++ app/src/main/java/com/example/android/politicalpreparedness/data/network/CivicsApiService.kt
@@ -44,7 +44,7 @@ interface CivicsApiService {
 
     @GET("voterinfo")
     suspend fun getVoterInfo(
-            @Query("electionId") id: Long,
+            @Query("electionId") id: Int,
             @Query("address") address: String, //address in format "state country".
             @Query("officialOnly") officialOnly: Boolean = false): VoterInfoResponse?
